Updated UserType
